LPS-95509 Extract asah logic from segments-service to segments-asah module
