+++ modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/configuration/SegmentsServiceConfiguration.java
@@ -32,9 +32,6 @@ import org.osgi.annotation.versioning.ProviderType;
 @ProviderType
 public interface SegmentsServiceConfiguration {
 
-	@Meta.AD(deflt = "15", name = "check-interval", required = false)
-	public int checkInterval();
-
 	@Meta.AD(deflt = "true", name = "segmentation-enabled", required = false)
 	public boolean segmentationEnabled();
